+++ package.json
@@ -60,7 +60,7 @@
     "typescript": "^5.0.2"
   },
   "peerDependencies": {
-    "vue": ">=3.3.0",
+    "vue": ">=3.5.0",
     "vue-router": "^4.1.6"
   },
   "devDependencies": {
