+++ src/components/Tabs/TabList.vue
@@ -1,0 +1,65 @@
+<template>
+  <TabList
+    class="relative flex items-center gap-7.5 overflow-x-auto border-b px-5"
+  >
+    <Tab
+      ref="tabRef"
+      as="template"
+      v-for="(tab, i) in tabs"
+      :key="i"
+      v-slot="{ selected }"
+      class="focus:outline-none focus:transition-none"
+    >
+      <slot v-bind="{ tab, selected }">
+        <button
+          class="flex items-center gap-1.5 border-b border-transparent py-3 text-base text-ink-gray-5 duration-300 ease-in-out hover:border-outline-gray-3 hover:text-ink-gray-9"
+          :class="{ 'text-ink-gray-9': selected }"
+        >
+          <component v-if="tab.icon" :is="tab.icon" class="size-4" />
+          {{ tab.label }}
+        </button>
+      </slot>
+    </Tab>
+    <div
+      ref="indicator"
+      class="tab-indicator absolute bottom-0 h-px bg-surface-gray-7"
+      :class="transitionClass"
+    />
+  </TabList>
+</template>
+<script setup>
+import { TabList, Tab } from '@headlessui/vue'
+import { ref, watch, computed, onMounted, nextTick, inject } from 'vue'
+
+const tabIndex = inject('tabIndex')
+const tabs = inject('tabs')
+
+const tabRef = ref([])
+const indicator = ref(null)
+const tabsLength = computed(() => tabs.value?.length)
+
+const transitionClass = ref('')
+
+function moveIndicator(index) {
+  if (index >= tabsLength.value) {
+    index = tabsLength.value - 1
+  }
+  const selectedTab = tabRef.value[index].el
+  indicator.value.style.width = `${selectedTab.offsetWidth}px`
+  indicator.value.style.left = `${selectedTab.offsetLeft}px`
+}
+
+watch(tabIndex, (index) => {
+  if (index >= tabsLength.value) {
+    tabIndex.value = tabsLength.value - 1
+  }
+  transitionClass.value = 'transition-all duration-300 ease-in-out'
+  nextTick(() => moveIndicator(index))
+})
+
+onMounted(() => {
+  nextTick(() => moveIndicator(tabIndex.value))
+  // Fix for indicator not moving on initial load
+  setTimeout(() => moveIndicator(tabIndex.value), 100)
+})
+</script>
