+++ src/index.js
@@ -33,6 +33,8 @@ export { default as Spinner } from './components/Spinner.vue'
 export { default as Switch } from './components/Switch.vue'
 export { default as TabButtons } from './components/TabButtons.vue'
 export { default as Tabs } from './components/Tabs/Tabs.vue'
+export { default as TabList } from './components/Tabs/TabList.vue'
+export { default as TabPanel } from './components/Tabs/TabPanel.vue'
 export { default as TextInput } from './components/TextInput.vue'
 export { default as Textarea } from './components/Textarea.vue'
 export {
